+++ .eslintrc.js
@@ -5,6 +5,7 @@ module.exports = {
     sourceType: 'module',
     ecmaVersion: '2017'
   },
+  parser: 'babel-eslint',
   settings: {
     polyfills: [
       'Array.from',
@@ -23,7 +24,16 @@ module.exports = {
   },
   overrides: [
     {
-      files: 'src/index-polyglot.js',
+      files: 'test/**',
+      env: {
+        mocha: true
+      },
+      globals: {
+        assert: true
+      }
+    },
+    {
+      files: 'src/index-node.mjs',
       globals: {
         require: true,
         __dirname: true
@@ -40,8 +50,15 @@ module.exports = {
       }
     },
     {
+      files: ['*.html'],
+      rules: {
+        'import/unambiguous': 0
+      }
+    },
+    {
       files: ['*.md'],
       globals: {
+        require: true,
         getJSON: true,
         args: true,
         url: true,
@@ -50,14 +67,16 @@ module.exports = {
       },
       rules: {
         'import/unambiguous': 0,
+        'import/no-commonjs': 0,
         'no-console': 0,
         'no-shadow': ['error', {
           allow: ['getJSON']
         }],
         'node/no-unsupported-features/es-builtins': 0,
         'node/no-unsupported-features/es-syntax': 0,
-        'promise/prefer-await-to-callbacks': 0,
         'node/no-missing-import': 0,
+        'node/no-missing-require': 0,
+        'promise/prefer-await-to-callbacks': 0,
         'padded-blocks': 0,
         'import/no-unresolved': 0,
         'promise/catch-or-return': 0,