+++ rollup.config.test.js
@@ -1,9 +1,0 @@
-// eslint-disable-next-line import/no-anonymous-default-export
-export default {
-  input: 'tests/test.js',
-  output: {
-    file: 'tests/test-node.js',
-    format: 'umd',
-    name: 'testJSON'
-  }
-};