+++ test/bootstrap/node.js
@@ -1,0 +1,3 @@
+import {buildGetJSON} from '../../src/index-node.mjs';
+
+global.buildGetJSON = buildGetJSON;